Map language clients to file extensions
Use MEF to create the language clients that implement the
ILanguageClient interface.

Use the IFileExtensionRegistryService to map a file extension to
a content type. The content type can then be used to find a
language client. The IFileExtensionRegistryService is provided by
the Microsoft.VisualStudio.CoreUtility assembly and maps
file name extensions to content type definitions for those
defined using FileExtensionToContentTypeDefinition.

Added a language server text editor extension which does not do
anything yet apart from try to find a language client for a file
extension.