Highlight diagnostic information in text editor
Handle the textdocument/publishDiagnostics message. Information
from this message is used to highlight text in the editor. Errors
and warnings are currently highlighted correctly. Hints and info
diagnostics are currently highlighted as warnings. A custom text
editor marker is needed for hints and information diagnostics as well
as a custom tooltip provider.