Show completion items
Send textDocument/completion message to language server when any
key is typed in the text editor. Completion items are then shown in
the text editor's completion list window.

More is required here to handle existing words when typing so the
correct characters are selected in the completion list window and
the correct characters are replaced when a completion list item is
selected.