Support full text sync language servers
The server capabilities are now checked to see if the language server
supports full or incremental document syncing. For full sync the
full document text is sent when the document has changed. If no
sync is supported then no textDocument/didChange message is sent.
Previously only incremental changes would be sent even if the
language server did not suppor them.