Implement go to declaration
Added the Go to declaration menu to the text editor's right click
menu. This sends a textDocument/definition message to the language
server. If no results are found then the search results window
is opened with no results. If a single result is found then that
is opened in the text editor. If multiple results are found then
the search results window is opened showing all the results.