Improve code completion window triggering
Trigger code completion on typing only when typing at the end of
the line. If a space is typed at the end of the line the code
completion window is not triggered. It is only triggered if the
last charactered typed was part of a word. Also apply the same
logic when the backspace key is pressed.

Previously code completion was triggered on every key press.