Support language clients with custom messages
Add ILanguageClientCustomMessage which a language client can
implement. This allows a custom target to be added to the JsonRpc
instance and also has an AttachForCustomMessageAsync method which
is passed the JsonRpc instance. The MiddleLayer property is not yet
supported.

Removed the implementations of AsyncEventHandler and InvokeAsync. These
are available from the Microsoft.VisualStudio.Threading NuGet
package. The StreamJsonRpc NuGet package has the Threading NuGet
package as a dependency. The custom implementations were causing
build errors if the language client needed to implement the
ILanguageClientCustomMessage since it has it uses JsonRpc and a
reference to the StreamJsonRpc would bring in the Threading NuGet
package as a dependency.