Support sending custom message to mock language server
Added a Language Server main menu with a Send Custom Message menu
item. This menu is shown when the mock language extension is used.
The Send Custom Message will send a message to the language server
using a custom message and show the message returned back from
the language server.