Use different TypeScript language server
Uses language server from:

https://github.com/sourcegraph/javascript-typescript-langserver

Does not work. The language client should be sending a Uri for the
file path.

Had to specify a root path in the parameters for the initialize
message since the language server will throw an error when it tries
to access a null projectManager when the root path is not defined.