Improve shutdown logic
Ensure JsonRpc is disposed if the initialize response has not yet
been returned. Ensure a new session is created if the last document
for a language client is being closed and a new document is opened.
Remove event handler for language client Started event before stopping
the client to prevent this handler being called.