Allow language clients to registered using a filename
Previously it was only possible to associated a language client with
a file extension. This does not work in all cases (e.g. Dockerfile).
Now the FileName attribute can be used to map a content type to
a filename. The IFileExtensionRegistryService2 supports finding
a content type based on filename or a filename extension. So this
allows a language client to be associated with a filename instead of
just a filename extension.

[Export]
[FileName ("myfile")] // Used instead of [FileExtension (".myext");
[ContentType ("MyContentType")]
internal static FileExtensionToContentTypeDefinition fileExtensionDefinition;