Show code completion when trigger character is typed
When a character is typed first check that it is in the list of
trigger characters defined by the language server and if so then
ask the language server for code completion items. Previously code
completion would only occur if a word was being typed and the caret
was at the end of the line.