Fix incorrect range for incremental text changes
When text was removed the end position for an incremental text change
sent to the language server in the textDocument/didChange was
incorrect. An attempt was made to get the position using the offset
in the TextChange after the text had changed. This could fail with
text being removed since the offset did not exist and result in a
position of -1,-1 being sent to the language server.