Add Yaml language client
Runs the https://github.com/redhat-developer/yaml-language-server
from the bin directory after it is installed using npm.

Currently this language client does not work. The settings sent
to the server seem to cause an error. The settings property
of the message seems to be null when it reaches the server.