Send workspace/didChangeConfiguration message from client
The workspace/didChangeConfiguration message is now sent to the
language server by the language client after the initialize message
has been sent. The configuration message is sent if a file called
VSWorkspaceSettings.json is found in the directory:

~/Library/Preferences/VisualStudio/7.0

The language client needs to define the ConfigurationSections and
only those that are defined there will be sent to the language
server.

Currently having a settings json file for a solution or folder is
not supported.