Support default VSWorkspaceSettings.json provided with language client
If a language client's assembly has a VSWorkspaceSettings.json file
embedded as a resource then that will be sent to the language server
after the initialize request has been sent. If there is a
VSWorkspaceSettings.json file in the user's preferences directory
(~/Library/Preferences/VisualStudio/7.0) then this information will
be merged into the default settings before sending the merged info
to the language server.