Switch to using content type as language client id
The language client id is used when logging to the Language Client
Log window. Previously this was the file extension but that can
be empty (e.g. Dockerfile). So now the content type name is used.
Also changed the logic in the LanguageClientSession's
IsSupportedDocument. This now uses the content type information for
a file instead of comparing the file extension. Using the file
extension may not match when it should, for example if a .yml and
a .yaml file are opened. In the Yaml file case the content type
is the correct match whilst a filename extension would not match
in all cases.