Support resolving completion items
When an item is selected in the code completion list window a
completionItem/resolve will be sent to the language server if it
supports completion resolve provider. The updated completion item
information will then be used in the description tooltip window
that appears next to the completion list window. Typically language
servers provide documentation in the response returned.