Fix incorrect text changes sent after document formatted
When a document formatting resulted in more than once change that
replaced characters the wrong range was sent for changes after the
first change. This was because the offset being used was the offset
before the text had changed and was incorrect based on the current
state of the text. Now the NewOffset is used which is the offset
based on the current text.