Use rename provider if language server supports it
If the language server provider indicates it supports renaming
then when a rename is triggered in the text editor then a dialog
is opened prompting for the new name for the item under the caret,
then a textDocument/rename message is sent to the language server.

Currently only renames in the open text editor are supported.

The rename dialog has a UI problem where as you type the first time
in the text box the controls move down slightly.