Fix TypeScript language server error when text changes sent
The TypeScript language server checks the rangeLength property of the
textDocument/didChange message and was throwing an exception since
this was non-zero. Now when the language server does not support
incremental text changes only the text is set on the didChange message
and the rangeLength is not set and will default to zero. Previously
the length of the full text was being set in the message.