Fix dot completion not showing any items
The word at the current caret was incorrectly taking the word before
the dot character and causing the completion list to be filtered.
This was because the offset was being used instead of the column
of the line. No word should be found at the current caret position
when the dot character is typed in.