Fix failure to re-launch TypeScript language client
The TypeScript language server does not seem to send a response
back to the exit message which causes the JsonRpc class to not be
disposed. To avoid this the exit message is sent with a timeout
so the JsonRpc connection will be disposed even if the server does
not respond. This then prevents the node process from not shutting
down and preventing a second launch of node again on closing and
re-opening a TypeScript file.