Handle renaming across files
When a language server does not support textDocument/rename but
supports textDocument/references the language client would use
text editor text links when the item being renamed existed in multiple
files. To mirror the behaviour the text editor for other languages,
if the item being renamed exists in multiple files the rename dialog
is displayed and then the new name is applied across the files.
If the item exists only in the current text editor then an inline
rename occurs using text links.