Fix exception when unknown content type open and tooltip requested
Opening a .cs file and then hovering would cause the language client
tooltip provider to run. The .cs file extension is not a known
content type provided by the language servers and would result in
an exception being thrown. Now a null is returned which is handled
by the language client tooltip provider.