Support running a quick fix/code action
Implemented workspace/executeCommand and workspace/applyEdits. This
allows a quick fix from the Quick Fixes menu to be run. Running the
quick fix will send a workspace/executeCommand to the language
server. The language server will then send a workspace/applyEdits
message back to the IDE and those text edits are applied to the
required documents.