Implement workspace symbol search
Searching in the quick search at the top right of the IDE will
now send a workspace/symbol message to the language server if it
is a workspace symbol provider. The results of this search will
appear in the search results with the category of Symbol. The
formatting of the search result matches the results when searching
in a C# project for types.

  symbolName
  symbolKind (file /path/to/file)

Clicking on the search result will open the corresponding file and
move the caret to search result location.