Add some support for snippets
Use MonoDevelop's CodeTemplate if a snippet completion item is
returned from the language server. This has some support for
TextMate snippets, which is what the language server protocol
snippets are based on.

For the Swift language server this does not work that well in
all cases. The Swift snippets use the same place holder name text
which results in the value being replaced for all arguments to
a function call when the arguments are for different parameters.