Improve snippet support for Swift language server
The Swift language server has snippets of the form:

    precondition(${1:value}, ${2:value})

This gets converted to the MonoDevelop snippet text:

    precondition($value$, $value$)

In Visual Studio Code these are not linked values they are different
arguments. In MonoDevelop these are linked so changing one changes
all of them. To match the Visual Studio Code behaviour the
conversion to a CodeTemplate now includes the number as the variable
name and uses the 'value' part as the default value for the variable.
This causes the text editor to show 'value' as part of the insert
snippet but does not link the text values. So the precondition
snippet text from the Swift language server is now converted to:

    precondition($value1$, $value2$)

Linked values are still supported by using the same number and name
in every location where the values should be linked, for example

    precondition(${1:value}, ${2:other}, ${1:value})