Fix didOpen message not being sent.
Previous code was using the Workbench.DocumentOpened event. The
Document from this event does not have the associated Editor set
which resulted in a null reference exception. Now the logic to
send didOpen and didClose has been moved to the language client's
DocumentControllerExtension. This extension uses the OnContentChanged
method to send the didOpen message once the TextEditor has been
associated with the document. If this happens before the language
client has started the language client's Started event is used to
send the didOpen message if it has not been sent.