Initialize language clients in document controller extension
Previously a startup handler would initialize the language clients.
Move this initialization to the document controller extension
to avoid the IDE logging a warning about MEF being used whilst
it is still being built. Originally the language client addin
would be the first to use the composition manager when a file was
opened and it would cause a long delay on opening the file. This no
longer seems to be the case when using the document controller
extension.