Fix exception on quitting IDE
An InvalidOperationException: Collection was modified.
error was thrown on quitting the IDE. The document
controller extension would shutdown the session
whilst the sessions were being enumerated and
shutdown on solution unload. There is no need to
shutdown the sessions on solution unload since the
controller extension handles this. ##enumeration operation may not execute.