Disable suggestion mode in new editor
When the language server is available for a file then suggestion
mode is disabled for the new editor. There does not seem to be
a simple way to provide suggestion mode support. It seems to work
for .cs files since they have a custom IAsyncCompletionItemManager
which behaves differently to the DefaultCompletionItemManager.
The custom completion item manager for .cs files sets the
selected item index to be 0 if there are no matches. However the
DefaultCompletionItemManager always sets the selected index to -1
if suggestion mode is enabled. This then prevents tab from
inserting the soft selected item in the completion list.

For now suggestion mode is disabled when a file has focus and
it has an associated language client.